% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nas.R
\name{check_nas}
\alias{check_nas}
\title{Check for missing values}
\usage{
check_nas(data, column)
}
\arguments{
\item{data}{tibble or data frame}

\item{column}{column to check for NAs}
}
\value{
stops if there are NAs in the column or if the column does not exist
}
\description{
This function checks if there are any NAs in a specified column of a data frame.
}
