% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_helpers.R
\name{create_a_matrix}
\alias{create_a_matrix}
\title{Create a matrix}
\usage{
create_a_matrix(post_summ)
}
\arguments{
\item{post_summ}{tibble with columns \code{variable} and \code{postmean},
variable examples are \code{a[1,1]}, \code{a[2,1]}, etc.}
}
\value{
matrix with rows and columns corresponding to the indices in the variable
}
\description{
This function takes a postsumm object and returns a matrix with the
relevant information
}
\keyword{internal}
