% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameter_subhierarchical.R
\name{extract_parameter_subhierarchical_simplified}
\alias{extract_parameter_subhierarchical_simplified}
\title{Title}
\usage{
extract_parameter_subhierarchical_simplified(
  hierarchical_data,
  subhierarchy,
  parname,
  fit_samples
)
}
\arguments{
\item{fit_samples}{}
}
\description{
Title
}
