% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(
  survey_df,
  y = "logit_indicator",
  area = "iso",
  area_select = NULL,
  runstep,
  global_fit = NULL,
  hierarchical_level = c("intercept", "subcluster", "iso"),
  add_subnational_hierarchy = "subnat",
  mu_isvector = FALSE,
  chains = 4,
  iter_sampling = 200,
  iter_warmup = 150,
  compile_model = TRUE,
  force_recompile = FALSE,
  seed = 1234,
  refresh = 200,
  adapt_delta = 0.9,
  max_treedepth = 14
)
}
\arguments{
\item{survey_df}{tibble with survey data}

\item{y}{column name of outcome.}

\item{area}{column name of the area of each observation (geounit; iso or subnational region)}

\item{area_select}{area name to use for local run (eg iso code or subnat region name)}

\item{runstep}{type of run, currently one of XXX (see Details).}

\item{global_fit}{optional object of class XX, used to obtain fixed
values to use for some parameters in the current fit (see Details).}

\item{hierarchical_level}{vector specifying hierarchical structure used for mu}

\item{add_subnational_hierarchy}{level that's added to the hierarchy for subnational, defaults to 'subnat'}

\item{mu_isvector}{Boolean, TRUE if mu is a vector, defaults to FALSE

Settings for sampling}

\item{chains}{number of chains to run}

\item{iter_sampling}{number of posterior samples to draw}

\item{iter_warmup}{number of warmup iterations}

\item{compile_model}{boolean indicator of whether to compile the Stan model}

\item{force_recompile}{boolean indicator of whether to force recompilation of the Stan model}

\item{seed}{random seed}

\item{refresh}{number of iterations between progress updates}

\item{adapt_delta}{target acceptance rate for the No-U-Turn Sampler}

\item{max_treedepth}{maximum tree depth for the No-U-Turn Sampler}
}
\value{
List that contains samples, stan_data, and other information relevant to model fit (arguments)
}
\description{
Fit model
}
\details{
The `fit_model` function fits the toy example for hierarchical models/seq fitting.
The argument \code{runstep} determines the type of run to perform. The
following run steps are supported:
- "step1a": Fit the global model
- "local_national": Fit the model to data from a single country, using a 1a fit.
- "global_subnational": Fit the model to global database with subnational data, using a 1a fit.
- "local_subnational": Fit the model to subnational data from a single country or region, using a global_subnational fit.
This is also explained in the documentation folder.

Details on hierarchical set ups used
The package allows the structure of the hierarchical prior to be configured by the user
through the  \code{hierarchical_level} argument.
These arguments expect a character vector that specifies a nesting hierarchical structure.
Each element of the vector must be either "intercept" or a column name in the survey dataset, where
"intercept" will add a global intercept for the parameter.
The vector must be in descending order in terms of the hierarchy: that is, it starts with
"intercept" and proceeds down the hierarchy.

For example, suppose we are fitting country-level data, where the dataset has columns
"name_country", "name_sub_region", and "name_region" containing the name of the country,
sub-region, and region that each observation belongs to. To specify that parameter mu
should be fitted with a hierarchical model in which countries are nested within sub-regions within regions within world,
we would use the argument
\code{hierarchical_level = c("intercept", "name_region", "name_sub_region", "name_country")}.

Optionally, model parameters can be fixed to values from a previous model fit
provided via the \code{global_fit} argument. In a typical use case, the
\code{global_fit} will have been fit to data from many geographic units
(e.g., all countries), while the current fit uses data from a smaller number
of locations.
When using a global_fit to fix parameter values, what exactly is fixed is determined by the runstep and global_fit combi.
see options in function
}
