% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_simplified.R
\name{fit_model_simplified}
\alias{fit_model_simplified}
\title{Simplified fitting function}
\usage{
fit_model_simplified(
  survey_df,
  y = "logit_indicator",
  area = "iso",
  area_select = NULL,
  runstep,
  global_fit = NULL,
  hierarchical_level = c("intercept", "subcluster", "iso"),
  add_subnational_hierarchy = "subnat",
  chains = 4,
  iter_sampling = 200,
  iter_warmup = 150,
  compile_model = TRUE,
  force_recompile = FALSE,
  seed = 1234,
  refresh = 200,
  adapt_delta = 0.9,
  max_treedepth = 14,
  stan_file_path = file.path(here::here("inst/stan/", "hierfunctions_seq.stan")),
  mu2param = FALSE
)
}
\arguments{
\item{survey_df}{tibble with survey data}

\item{y}{column name of outcome.}

\item{area}{column name of the area of each observation (geounit, iso or subnational region)}

\item{area_select}{area name to use for local run (eg iso code or subnat region name)}

\item{runstep}{type of run, currently one of XXX (see Details).}

\item{global_fit}{optional object of class XX, used to obtain fixed
values to use for some parameters in the current fit (see Details).}

\item{hierarchical_level}{vector specifying hierarchical structure used for mu

Settings for sampling}

\item{chains}{number of chains to run}

\item{iter_sampling}{number of posterior samples to draw}

\item{iter_warmup}{number of warmup iterations}

\item{compile_model}{boolean indicator of whether to compile the Stan model}

\item{force_recompile}{boolean indicator of whether to force recompilation of the Stan model}

\item{seed}{random seed}

\item{refresh}{number of iterations between progress updates}

\item{adapt_delta}{target acceptance rate for the No-U-Turn Sampler}

\item{max_treedepth}{maximum tree depth for the No-U-Turn Sampler}
}
\description{
Simplified fitting function
}
\details{
The `fit_model_simplified` function fits the toy example for hierarchical models/seq fitting.
The argument \code{runstep} determines the type of run to perform. The
following run steps are supported:
- "step1a": Fit the global model
- "local_national": Fit the model to data from a single country, using a 1a fit.
This is also explained in the documentation folder.

Details on hierarchical set ups used
Several area-specific parameters of the fpemplus model have hierarchical priors
assigned to them so that information can be shared between areas.
The package allows the structure of the hierarchical prior to be configured by the user
through the  \code{hierarchical_level} argument.
These arguments expect a character vector that specifies a nesting hierarchical structure.
Each element of the vector must be either "intercept" or a column name in the dataset, where
"intercept" will add a global intercept for the parameter.
The vector must be in descending order in terms of the hierarchy: that is, it starts with
"intercept" and proceeds down the hierarchy.

For example, suppose we are fitting country-level data, where the dataset has columns
"name_country", "name_sub_region", and "name_region" containing the name of the country,
sub-region, and region that each observation belongs to. To specify that the spline coefficients
should be fitted with a hierarchical model in which countries are nested within sub-regions within regions within world,
we would use the argument
\code{hierarchical_splines = c("intercept", "name_region", "name_sub_region", "name_country")}.

Optionally, model parameters can be fixed to values from a previous model fit
provided via the \code{global_fit} argument. In a typical use case, the
\code{global_fit} will have been fit to data from many geographic units
(e.g., all countries), while the current fit uses data from a smaller number
of locations.
Note: remainder of details is currently determined by the runstep
To use a global fit to fix parameter values, a number of settings must be the
same in the global fit and the current call to `fpemplus`:
- For any of the hierarchical settings, e.g.
  \code{hierarchical_asymptote}, all of the hierarchical levels
  used for that parameter in the global fit must also be used in the current
  fit. For instance, if the `global_fit` used
  \code{hierarchical_asymptote = c("intercept", "name_region", "name_country")},
  then in the current fit it is valid to use
  \code{hierarchical_asymptote = c("intercept", "name_region", "name_country")} again or
  \code{hierarchical_asymptote = c("intercept", "name_region", "name_country", "name_subnational")},
  but it is not valid to use
  \code{hierarchical_asymptote = c("intercept", "name_region", "name_subnational")}.
- All hierarchical levels for terms and sigmas to fix in the current fit are
  contained in the levels in the hierarchy used for the corresponding
  parameters in the global fit. For example, if
  \code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_region", "name_country")}
  then the global fit must have included at least "intercept", "name_region",
  and "name_country" for \code{hierarchical_asymptote}.
- Any hierarchical levels to fix in the current fit are at the highest levels
  of the hierarchical structure. For example, if
  \code{hierarchical_asymptote = c("intercept", "name_region", "name_country")},
  then it is valid to use
  \code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_region")},
  but it is not valid to use
  \code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_country")}.
- It is not valid to fix terms at a given hierarchy level without also fixing
  the sigma estimate at that hierarchy level. For example, we cannot specify
  \code{hierarchical_asymptote_sigmas_fixed = c("intercept")} and
  \code{hierarchical_asymptote_terms_fixed = c("intercept", "name_region")}
- It is only valid to set `fix_smoothing = TRUE` if also `smoothing = TRUE`.
- All settings for the arguments `model`, `t_star`, `smoothing`, `tau_prior`,
  and `rho_prior` must be the same in the `global_fit` and the current fit.
- If `hierarchical_splines_sigmas_fixed` or
  `hierarchical_splines_terms_fixed` include any hierarchical levels (i.e.,
  if either is different from the empty vector `c()`), all settings for
  `num_knots` and `spline_degree` must be the same in the `global_fit` and
  the current fit.
- All geographic units that appear in the `data` for the current fit at
  hierarchical levels for which any parameter is fixed must have also been
  included in the `data` used for the `global_fit`.
- If `fix_nonse = TRUE`, all data `source`s that appear in the `data` for the
  current fit must also have been included in the `data` used for the
  `global_fit`.
}
