% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_unit_index_data.R
\name{get_geo_unit_index_data}
\alias{get_geo_unit_index_data}
\title{Get unique hierarchical levels from survey data}
\usage{
get_geo_unit_index_data(data, hierarchical_levels, area)
}
\arguments{
\item{data}{survey data of interest that contains the hierarchical_column_names}

\item{hierarchical_levels}{vector that contains one of more specs of hierarchical_level (incl intercept)}

\item{area}{unit of analysis (eg country or subnational region)}
}
\value{
tibble with unique hierarchical levels (one column for each level) and index "c" assigned based on area
}
\description{
Get unique hierarchical levels from survey data and assign index "c" based on area.
}
\examples{
survey_dat <- tibble::tibble(subcluster = c("A", "A", "B", "B"),
    iso = c("iso1", "iso2", "iso3", "iso4"))
get_geo_unit_index_data(survey_dat,
        hierarchical_levels = c("intercept", "subcluster", "iso"),
        area = "iso")
}
