% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_raw.R
\name{get_mu_raw_labeled}
\alias{get_mu_raw_labeled}
\title{Title}
\usage{
get_mu_raw_labeled(fit, parname, morethan1param = FALSE)
}
\arguments{
\item{fit}{needs to include parname_raw_estimate}

\item{parname}{selected parameter name (example: "mu")}

\item{morethan1param}{does paramname refer to more than 1 parameter (a vector)}
}
\value{
tibble with samples of parname_raw_estimate, including info on hierarchical level
}
\description{
Title
}
