% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summaries.R
\name{get_posterior_summaries}
\alias{get_posterior_summaries}
\title{get_posterior_summaries}
\usage{
get_posterior_summaries(fit, params = c("mu_raw", "mu_sigma", "nonse"))
}
\arguments{
\item{fit}{fit object to summarize}

\item{params}{vector of parnames to summarize (w/o index)}
}
\value{
tibble with variable (parname with index used in model) and posterior mean
}
\description{
get_posterior_summaries
}
