% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summaries.R
\name{get_posterior_summaries_localhierarchy}
\alias{get_posterior_summaries_localhierarchy}
\title{Posterior summaries for use in local models}
\usage{
get_posterior_summaries_localhierarchy(
  fit,
  params = c("mu_raw", "mu_sigma", "nonse")
)
}
\arguments{
\item{fit}{fit object to summarize}

\item{params}{vector of parnames to summarize (w/o index)}
}
\value{
tibble with variable (parname with index used in model) and postmean (posterior mean)
}
\description{
Get posterior summaries for a set of parameters for use in local model
}
