% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summaries.R
\name{get_posterior_summaries_simplified}
\alias{get_posterior_summaries_simplified}
\title{get_posterior_summaries_simplified}
\usage{
get_posterior_summaries_simplified(
  fit,
  process_indicator_prefixes = c(""),
  dm_indicator_prefixes = c(""),
  process_params = c("mu_raw", "mu_sigma"),
  dm_params = c("nonse")
)
}
\arguments{
\item{fit}{# fit object to summarize}

\item{process_indicator_prefixes}{# prefixes for parameters from process model}

\item{dm_params}{}
}
\description{
get_posterior_summaries_simplified
}
