% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summary_one_fixable_param.R
\name{get_posterior_summary_one_fixable_param}
\alias{get_posterior_summary_one_fixable_param}
\title{Title}
\usage{
get_posterior_summary_one_fixable_param(fit, prefixed_param_name)
}
\arguments{
\item{fit}{fit object}

\item{prefixed_param_name}{parameter name, can include an index}
}
\description{
Title
}
