% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summaries.R
\name{get_posterior_summary_one_param}
\alias{get_posterior_summary_one_param}
\title{get_posterior_summary_one_param}
\usage{
get_posterior_summary_one_param(fit, param_name)
}
\arguments{
\item{fit}{fit object}

\item{param_name}{parameter name w/o index

need to get the estimate and/or fixed value for each indexed parameter
difficulty = stacking estimate on top of fixed values (for vector or array)
note that this does NOT get into hierarchical info (and is same for hier and nonhier param)
it is just the stacking}
}
\value{
tibble with variable (name with index) and posterior mean
}
\description{
get_posterior_summary_one_param
}
