% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{hierarchical_data}
\alias{hierarchical_data}
\title{Hierarchical_data}
\usage{
hierarchical_data(geo_unit_index_data, hierarchical_level)
}
\arguments{
\item{geo_unit_index_data}{tibble with unique hierarchical levels (one column for each level)}

\item{hierarchical_level}{vector with hierarchical spec}
}
\value{
list with model matrix and info on start and end indices associated with different levels
}
\description{
Create hierarchical data structure for a hierarchical model.
}
