% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_model_matrix.R
\name{hierarchical_model_matrix}
\alias{hierarchical_model_matrix}
\title{hierarchical_model_matrix
Create list with information including a model matrix, associated with the geo_unit_data and hierarchical_levels}
\usage{
hierarchical_model_matrix(geo_unit_index_data, hierarchical_level)
}
\arguments{
\item{geo_unit_index_data}{tibble with unique hierarchical levels (one column for each level, one row per combi)}

\item{hierarchical_level}{vector with hierarchical spec from highest to lowest levels}
}
\value{
list (modelmatrix, assign, index) associated with the geo_unit_data and hierarchical_levels
in modelmatrix each column refers to one eta (and nrows is the number of lowest level units).
assign is length of etas, with 0 for intercept, then an hierarchical level index (starting at 2) for each eta level
index is a tibble with n_eta rows, with info for each eta in and columns i [index of eta], column [hierarchical level], and level [name of hierarchical level]
}
\description{
hierarchical_model_matrix
Create list with information including a model matrix, associated with the geo_unit_data and hierarchical_levels
}
