% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_fun_localhierarchy}
\alias{init_fun_localhierarchy}
\title{Set initial values}
\usage{
init_fun_localhierarchy(chain_id, stan_data)
}
\arguments{
\item{chain_id}{the chain ID, used for setting the seed}

\item{stan_data}{a list containing the Stan data, to obtain length of parameter vectors}
}
\value{
a list of initial values for the model parameters
}
\description{
This function sets the initial values for the parameters in a simple hierarchical model.
}
