% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_helpers.R
\name{is_there_a_comma}
\alias{is_there_a_comma}
\title{Is there a comma?}
\usage{
is_there_a_comma(variable)
}
\arguments{
\item{variable}{character string to be checked for comma}
}
\value{
Boolean
}
\description{
Is there a comma?
}
\keyword{internal}
