% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalHierarchy-package.R
\docType{package}
\name{localhierarchy-package}
\alias{localhierarchy}
\alias{localhierarchy-package}
\title{localhierarchy: An R package to facilitate fitting of global and local Bayesian hierarchical models}
\description{
The localhierarchy package implements a flexible set up to estimate parameters hierarchically and to carry out local model fits, whereby higher-level parameters are fixed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlkemaLab/localhierarchy}
  \item \url{https://alkemalab.github.io/localhierarchy/}
  \item Report bugs at \url{https://github.com/AlkemaLab/localhierarchy/issues}
}

}
\author{
\strong{Maintainer}: Leontine Alkema \email{lalkema@umass.edu}

Authors:
\itemize{
  \item Shauna Mooney
  \item Evan Ray
  \item Herbert Susmann \email{herbps10@gmail.com}
}

}
\keyword{internal}
