% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_mu_raw}
\alias{plot_mu_raw}
\title{Plot prior and posterior densities of mu_raw}
\usage{
plot_mu_raw(fit, parname, morethan1param = FALSE, nresultsperpage = 30)
}
\arguments{
\item{fit}{List that includes "parname"_raw_estimate and stan_data}

\item{parname}{Selected parameter name (example: mu)}

\item{morethan1param}{Logical, does parname refer to more than 1 parameter (a vector)}

\item{nresultsperpage}{Number of results per page in summary plots}
}
\value{
lists with list 'summary_plots' and list 'plots_allmuraw'.
summary plots gives summary CIs, nresultsperpage at a time.
plots_allmuraw gives all the individual plots of each mu_raw, with density per chain and prior added
If morethan1param = TRUE, then each list contains a list per parameter k
#' @export
}
\description{
This function plots the posterior densities of mu_raw parameters, with prior added.
}
