% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_summaries.R
\name{plot_posterior_summaries}
\alias{plot_posterior_summaries}
\title{Plot summaries of hierarchical parameters}
\usage{
plot_posterior_summaries(
  res,
  hierarchy_select = NULL,
  areas_select = NULL,
  res2 = NULL,
  modelname1 = "model 1",
  modelname2 = "model 2",
  k_select = NULL,
  dodge = position_dodge(width = 0.5)
)
}
\arguments{
\item{res}{output from posterior_summary_hierparam}

\item{hierarchy_select}{optional, what hierarchical level to show? if NULL, all levels are shown}

\item{areas_select}{optional: areas to filter by, allowed only if one hierarchical level is selected}

\item{res2}{optional, output from posterior_summary_hierparam}

\item{modelname1}{label for res}

\item{modelname2}{label for res2}

\item{k_select}{optional, if res contains k, which k values to show? if NULL, all k values are shown}

\item{dodge}{used for offsetting plots, default is 0.5}
}
\value{
ggplot object
}
\description{
Display outputs from posterior_summary_hierparam
}
