% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_prior_post_sigmas_localhierarchy}
\alias{plot_prior_post_sigmas_localhierarchy}
\title{Plot prior and posteriors of sigmas from hierarchical models}
\usage{
plot_prior_post_sigmas_localhierarchy(fit, parname)
}
\arguments{
\item{fit}{List, needs to include parname_sigma_estimate and stan_data}

\item{parname}{Selected parameter name (example: "mu")}
}
\value{
Plot with density of sigma_estimate and prior added
}
\description{
This function plots the prior and posterior densities of sigma_estimate parameters,
}
