% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary_hierparam.R
\name{posterior_summary_hierparam}
\alias{posterior_summary_hierparam}
\title{posterior_summary_hierparam}
\usage{
posterior_summary_hierparam(
  fit,
  parname,
  morethan1param = FALSE,
  hierarchical_levels = fit$hierarchical_level
)
}
\arguments{
\item{fit}{needs to include parname_star}

\item{parname}{selected parameter name (example: "mu")}

\item{morethan1param}{does paramname refer to more than 1 parameter (a vector)}
}
\value{
list with summaries of mu for each hierarchical level (units with each level)
 these mus are obtained by summing up all relevant etas
for morethan1param, k refers to the index
}
\description{
posterior_summary_hierparam
}
