% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_helpers.R
\name{split_var_name_and_index}
\alias{split_var_name_and_index}
\title{split_var_name_and_index
For a data frame of estimates with a column that contains combined info
about a variable name and row and/or column indices, add columns containing
just the variable name and the indices.}
\usage{
split_var_name_and_index(estimate_df, num_inds)
}
\arguments{
\item{estimate_df}{data frame with a column called \code{variable} that contains \code{parname[i]} or \code{parname[i,j]} entries}

\item{num_inds}{number of indices in the variable name (used for 1 or 2)}
}
\value{
data frame with the same columns as \code{estimate_df} plus columns containing
just the variable name and the indices.
}
\description{
split_var_name_and_index
For a data frame of estimates with a column that contains combined info
about a variable name and row and/or column indices, add columns containing
just the variable name and the indices.
}
