% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary_hierparam.R
\name{summary_df}
\alias{summary_df}
\title{Compute summaries}
\usage{
summary_df(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
tibble with 2.5th quantile, posterior mean, and 97.5th quantile, and quant = c(0.025, 0.5, 0.975)
}
\description{
This function computes posterior summaries for a given vector and returns them in a tibble format.
}
