% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_helpers.R
\name{variable_exists_in_draws}
\alias{variable_exists_in_draws}
\title{Determine whether or not a variable exists in draws from a fit
This is similar to the function posterior:::check_existing_variables,
but note that function is not exported from posterior}
\usage{
variable_exists_in_draws(fit, var_name)
}
\arguments{
\item{fit}{fpemplus fit object}

\item{var_name}{name of variable as character string}
}
\value{
TRUE or FALSE indicating whether or not the variable exists in the
fit.
}
\description{
Determine whether or not a variable exists in draws from a fit
This is similar to the function posterior:::check_existing_variables,
but note that function is not exported from posterior
}
