% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary_hierparam.R
\name{extract_parameter_subhierarchical}
\alias{extract_parameter_subhierarchical}
\title{extract_parameter_subhierarchical}
\usage{
extract_parameter_subhierarchical(
  hierarchical_data,
  subhierarchy,
  parname,
  star_samples,
  morethan1param = FALSE,
  add_standardizedmu = FALSE
)
}
\arguments{
\item{hierarchical_data}{hierarchical_data, obtained from hierarchical_data(fit$geo_unit, fit$hierarchical_level)}

\item{subhierarchy}{selected hierarchical_level (example: "intercept" or "region")}

\item{parname}{selected parameter name (example: "mu")}

\item{star_samples}{samples from fit$samples for parname_star}

\item{morethan1param}{does parname refer to more than 1 parameter (a vector)}
}
\value{
a tibble with posterior samples of the selected parameter mu at the subhierarchy level.
these mus are obtained by summing up all relevant etas
variables are \code{name}, \code{value}, \verb{.draw/iteration/chain' and }k` referring to parameter index if morethan1param = TRUE
}
\description{
extract_parameter_subhierarchical
}
\keyword{internal}
