% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_model_matrix.R
\name{hierarchical_model_matrix}
\alias{hierarchical_model_matrix}
\title{hierarchical_model_matrix}
\usage{
hierarchical_model_matrix(geo_unit_index_data, hierarchical_level)
}
\arguments{
\item{geo_unit_index_data}{tibble with unique hierarchical levels (one column for each level, one row per combi)}

\item{hierarchical_level}{vector with hierarchical spec from highest to lowest levels}
}
\value{
A list with components \code{modelmatrix}, \code{assign}, and \code{index}:
\itemize{
\item \code{modelmatrix}: A matrix where each column refers to one eta (number of rows equals number of lowest level units).
\item \code{assign}: Integer vector of length equal to the number of etas; \code{0} for the intercept, then a hierarchical level index (starting at 2) for each eta level.
\item \code{index}: A tibble with \code{n_eta} rows, and columns:
\itemize{
\item \code{i}: Index of eta
\item \code{column}: Hierarchical level
\item \code{level}: Name of hierarchical level
}
}

All components are associated with the provided \code{geo_unit_data} and \code{hierarchical_levels}.
}
\description{
Create list with information including a model matrix, associated with the geo_unit_data and hierarchical_levels
}
