% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalHierarchy-package.R
\docType{package}
\name{localhierarchy-package}
\alias{localhierarchy}
\alias{localhierarchy-package}
\title{localhierarchy: Flexible set up for hierarchical modeling and local fitting}
\description{
The LocalHierarchy package implements a flexible set up to estimate parameters hierarchically and to carry out local model fits, whereby higher-level parameters are fixed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alkemalab.github.io/localhierarchy/}
}

}
\author{
\strong{Maintainer}: Leontine Alkema \email{lalkema@umass.edu}

Authors:
\itemize{
  \item Evan Ray
  \item Herbert Susmann \email{herbps10@gmail.com}
}

}
\keyword{internal}
