% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_mu_raw}
\alias{plot_mu_raw}
\title{Plot estimates of mu raws}
\usage{
plot_mu_raw(fit, parname, morethan1param = FALSE, nresultsperpage = 30)
}
\arguments{
\item{fit}{needs to include parname_raw_estimate and stan_data}

\item{parname}{selected parameter name (example: "mu")}

\item{morethan1param}{does paramname refer to more than 1 parameter (a vector)}

\item{nresultsperpage}{number of results per page in summary plots}
}
\value{
lists with list 'summary_plots' and list 'plots_allmuraw'.
summary plots gives summary CIs, nresultsperpage at a time.
plots_allmuraw gives all the individual plots of each mu_raw, with density per chain and prior added
If morethan1param = TRUE, then each list contains a list per parameter k
#' @export
}
\description{
Plot estimates of mu raws
}
