% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_prior_post_sigmas}
\alias{plot_prior_post_sigmas}
\title{plot_prior_post_sigmas}
\usage{
plot_prior_post_sigmas(fit, parname)
}
\arguments{
\item{fit}{needs to include parname_sigma_estimate and stan_data}

\item{parname}{selected parameter name (example: "mu")}
}
\value{
plot with density of sigma_estimate and prior added
}
\description{
plot_prior_post_sigmas
}
