% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_helpers.R
\name{split_var_name_and_index}
\alias{split_var_name_and_index}
\title{For a data frame of estimates with a column that contains combined info
about a variable name and row and/or column indices, add columns containing
just the variable name and the indices.}
\usage{
split_var_name_and_index(estimate_df, num_inds)
}
\description{
For a data frame of estimates with a column that contains combined info
about a variable name and row and/or column indices, add columns containing
just the variable name and the indices.
}
